implement cmd_craft_deliver

+++ src/server/platforms/ssdstm/endpoints/main/commands/cmd_craft_deliver.hpp
@@ -7,19 +7,19 @@ namespace emulator::ssd
 	{
         struct param_t
         {
-            std::uint32_t bad_status_1_risk;
-            std::uint32_t bad_status_2_risk;
-            std::uint32_t bad_status_3_risk;
-            std::uint32_t bad_status_4_risk;
-            std::uint32_t clean_water;
-            std::uint32_t food;
-            std::uint32_t inventory_index;
-            std::uint32_t inventory_index2;
-            std::uint32_t inventory_type;
-            std::uint32_t inventory_type2;
-            std::uint32_t list_type;
-            std::uint32_t medical_supplies;
-            std::uint32_t num;
+            std::int16_t bad_status_1_risk;
+            std::int16_t bad_status_2_risk;
+            std::int16_t bad_status_3_risk;
+            std::int16_t bad_status_4_risk;
+            std::int16_t clean_water;
+            std::int16_t food;
+            std::uint16_t inventory_index;
+            std::uint16_t inventory_index2;
+            std::uint8_t inventory_type;
+            std::uint8_t inventory_type2;
+            std::uint8_t list_type;
+            std::int16_t medical_supplies;
+            std::uint16_t num;
         };
 
 		json::value execute(json::value& data, const std::optional<database::users::user>& user) override;
